requirements.txt
langchain==0.2.3
langchain_openai==0.1.8
python-dotenv==1.0.1